feat(db): allow array workloads
Previously, workloads with arrays (as strings) would get them stored as
strings. Arrays in the DB, however, would fail to be converted properly
when converted from Mongo to CronJob, so the data would get dropped.

This allows us to properly save/restore arrays, and to view them in the
UI.